# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest API for TT APP",
        "title": "TT Rest API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/scanDetails": {
            "get": {
                "description": "Retrieves scan results filtered by a customer ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scans"
                ],
                "summary": "Get scan details by scan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan ID",
                        "name": "scanID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Scan"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/scansByCustomerID": {
            "get": {
                "description": "Retrieves scan results filtered by a customer ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scans"
                ],
                "summary": "Get scans by customer ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer ID",
                        "name": "customerID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Scan"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Scan": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                },
                "resourceName": {
                    "type": "string"
                },
                "resourceType": {
                    "type": "string"
                },
                "riskScore": {
                    "type": "integer"
                },
                "tags": {
                    "type": "string"
                }
            }
        }
    }
}